test: refactor locations of test code
